requirements.txt
pytest
selenium
percy-selenium